// idfistafor/src/app/shared/models/trasformazione.model.ts
/*!
 * SPDX-FileCopyrightText: (C) Copyright 2023 Regione Piemonte
 * 
 * SPDX-License-Identifier: EUPL-1.2
 */
export class Trasformazione {
	idIstanza?: number;
	nrIstanza?: number;
	anno?: string;
	statoIstanza?: string;
	richiedente?: string;
	comune?: string;
	compensazione?: string;

	displayItem?: string;
}